Large refactor to better deal with objects
After getting base idea flushed, now we needed to organize the code so it is more useable.

Add a shape class heirachy to manage movement of shapes and vectors

Added a screen class to free up main and handle the main Game/Sim loop

Break the algorithms out into there own classes, static as of now, to adjust later as needed.